Work on color space support

+++ src/ColorMapModel.cpp
@@ -121,7 +121,7 @@ QVariant ColorMapModel::data(const QModelIndex& index, int role /* = Qt::Display
 		}
 
 		case Qt::TextAlignmentRole:
-			return Qt::AlignLeft | Qt::AlignVCenter;
+			return static_cast<int>(Qt::AlignLeft | Qt::AlignVCenter);
 
 		default:
 			break;
